Add add-source for opentelemetry
